Move raiden/store/state to raiden/state
