+++ raiden/tests/unit/epics.spec.ts
@@ -32,7 +32,7 @@ import fetch from 'cross-fetch';
 import { ShutdownReason } from 'raiden/constants';
 import { RaidenAction } from 'raiden/actions';
 import { raidenReducer } from 'raiden/reducer';
-import { RaidenState, initialState } from 'raiden/store/state';
+import { RaidenState, initialState } from 'raiden/state';
 import { raidenInit, raidenShutdown } from 'raiden/store/actions';
 import {
   newBlock,
