+++ raiden/tests/unit/mocks.ts
@@ -56,7 +56,7 @@ import TokenAbi from 'raiden/abi/Token.json';
 
 import { RaidenEpicDeps } from 'raiden/types';
 import { RaidenAction } from 'raiden/actions';
-import { RaidenState, initialState } from 'raiden/store/state';
+import { RaidenState, initialState } from 'raiden/state';
 import { Address, Signature } from 'raiden/utils/types';
 
 type MockedContract<T extends Contract> = jest.Mocked<T> & {
@@ -75,14 +75,17 @@ interface MockRaidenEpicDeps extends RaidenEpicDeps {
   getTokenContract: (address: string) => MockedContract<Token>;
 }
 
+/**
+ * Create a mock of RaidenEpicDeps
+ *
+ * @returns  Mocked RaidenEpicDeps
+ */
 export function raidenEpicDeps(): MockRaidenEpicDeps {
   const network: Network = { name: 'testnet', chainId: 1337 };
 
   const provider = new JsonRpcProvider() as jest.Mocked<JsonRpcProvider>;
 
-  /**
-   * spyOn .on, .removeListener & .emit methods and replace with a synchronous simplified logic
-   */
+  // spyOn .on, .removeListener & .emit methods and replace with a synchronous simplified logic
   const mockEthersEventEmitter = (target: JsonRpcProvider | Contract): void => {
     const listeners = new Map<EventType, Set<Listener>>();
     function getEventTag(event: EventType): string {
@@ -188,6 +191,12 @@ export function raidenEpicDeps(): MockRaidenEpicDeps {
   };
 }
 
+/**
+ * Create a mocked ethers Log object
+ *
+ * @param filter  EventFilter object
+ * @returns  Log object
+ */
 export function makeLog({ filter, ...opts }: { filter: EventFilter } & Partial<Log>): Log {
   const blockNumber = opts.blockNumber || 1337;
   return {
@@ -205,7 +214,13 @@ export function makeLog({ filter, ...opts }: { filter: EventFilter } & Partial<L
   };
 }
 
-/* Returns a mocked MatrixClient */
+/**
+ * Returns a mocked MatrixClient
+ *
+ * @param userId  userId of account owner
+ * @param server  server mock hostname
+ * @returns  Mocked MatrixClient
+ */
 export function makeMatrix(userId: string, server: string): jest.Mocked<MatrixClient> {
   return (Object.assign(new EventEmitter(), {
     startClient: jest.fn(async () => true),
@@ -247,7 +262,11 @@ export function makeMatrix(userId: string, server: string): jest.Mocked<MatrixCl
   }) as unknown) as jest.Mocked<MatrixClient>;
 }
 
-/* Returns some valid signature */
+/**
+ * Returns some valid signature
+ *
+ * @returns  Some arbitrary valid signature hex string
+ */
 export function makeSignature(): Signature {
   return '0x5770d597b270ad9d1225c901b1ef6bfd8782b15d7541379619c5dae02c5c03c1196291b042a4fea9dbddcb1c6bcd2a5ee19180e8dc881c2e9298757e84ad190b1c' as Signature;
 }
